requirements.txt
# Kafka dependencies
kafka-python==2.0.2
confluent-kafka==2.0.2

# Data processing
pyspark==3.3.1
pymongo==4.3.3
pandas==1.5.3
numpy==1.24.2

# AWS integration
boto3==1.26.84
s3fs==2023.3.0

# Utilities
python-dotenv==1.0.0
jsonschema==4.17.3

# API and dashboard
flask==2.2.3
dash==2.9.1
plotly==5.13.1

# Testing
pytest==7.3.1